feat: Update Product and User services to use Integer for ID parameters

+++ src/main/java/swd392/backend/domain/controller/ProductController.java
@@ -19,7 +19,7 @@ public class ProductController {
     }
 
     @GetMapping("/{id}")
-    public ProductDTO findById(@PathVariable Long id) {
+    public ProductDTO findById(@PathVariable Integer id) {
         return productService.findProductById(id);
     }
 
@@ -34,7 +34,7 @@ public class ProductController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteProductById(@PathVariable Long id) {
+    public void deleteProductById(@PathVariable Integer id) {
         productService.deleteProduct(id);
     }
 }
