+++ src/main/java/swd392/backend/domain/service/user/UserServiceImpl.java
@@ -23,7 +23,7 @@ public class UserServiceImpl implements UserService{
     }
 
     @Override
-    public UserDTO findUserById(Long id) {
+    public UserDTO findUserById(Integer id) {
         return userRepository.findById(id).map(userMapper::toDto).orElse(null);
     }
 }
